requirements/requirements_test.txt
# pre commit hooks
pre-commit
importlib-metadata
importlib-resources
zipp
six
identify
aspy
aspy.yaml
pyyaml
isodate
cfgv
toml

# code formatting
black

# Factories
factory_boy==2.12.0
faker==2.0.0
text-unidecode==1.2

# pytest
py==1.8.0
pytest==4.4.1
pytest-django==3.4.8
pytest-env==0.6.2
pytest-socket==0.3.3
atomicwrites==1.3.0
pluggy==0.10.0
more-itertools==7.0.0
python-dateutil==2.8.0
attrs==19.1.0

# test coverage
pytest-cov==2.7.1
coverage==4.5.3
docopt==0.6.2
